backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.44.0

# HTTP requests for HuggingFace API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0
